Function for creating dataset with given random values for specified variables

Function for creating dataset with given random values for specified variables

Usage

dsR(vr = "ms:3;bi:4;ii:1", nr = 100)

Arguments

vr

Array of variables with types for creating a data frame.

nr

Number of rows for given dataframe!

Value

dataframe

Examples

dsR(vr="ms:3;bi:4;ii:1", nr=100);
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#>       ms_1   ms_2         ms_3 bi_1 bi_2 bi_3 bi_4 ii_1
#> 1    20123 102023        Tokyo    1    1   No    1  614
#> 2    54903   3502 Antananarivo    1    0   No    1  589
#> 3    20123   3502        Tokyo    1    1  Yes    1   17
#> 4   102023 102023    Ljubljana    1    1   No    0  562
#> 5     3502   3502    Ljubljana    1    0   No    0  199
#> 6     3502  54903      Seattle    0    0   No    1  731
#> 7    20123  20123       Taipeh    0    1   No    0  743
#> 8    54903 102023      Seattle    0    1  Yes    0  953
#> 9   102023  20123 Antananarivo    1    1  Yes    0  158
#> 10  102023  20123       Taipeh    1    0  Yes    0  392
#> 11   54903  54903    Ljubljana    1    0   No    1  156
#> 12  102023 102023     Windhoek    1    0   No    0  975
#> 13    1000   1000      Seattle    1    1   No    0  529
#> 14    1000  20123       Taipeh    0    1  Yes    1  159
#> 15   20123   1000    Ljubljana    0    1  Yes    1   77
#> 16   54903  20123      Seattle    0    0   No    0  581
#> 17   20123  54903       Taipeh    0    1  Yes    1  760
#> 18    3502 102023 Antananarivo    1    0   No    1  452
#> 19  102023   1000     Asuncion    0    0  Yes    1  171
#> 20   20123   1000     Asuncion    1    0   No    0   31
#> 21   20123  54903    Ljubljana    0    0   No    1  915
#> 22    3502  54903       Taipeh    0    1  Yes    0  756
#> 23  102023 102023 Antananarivo    0    0   No    0  803
#> 24  102023  54903 Antananarivo    1    0  Yes    1  262
#> 25    1000   1000 Antananarivo    1    1   No    1  472
#> 26    1000  20123       Taipeh    1    0   No    0  690
#> 27  102023 102023     Windhoek    0    0   No    0  633
#> 28   20123  20123      Seattle    0    1   No    1  976
#> 29    3502 102023        Tokyo    0    1   No    0  338
#> 30   54903   1000       Taipeh    0    0  Yes    0  280
#> 31    1000 102023    Ljubljana    1    1   No    1  383
#> 32  102023   1000     Asuncion    1    1  Yes    0  391
#> 33    3502   3502     Windhoek    1    0   No    0  451
#> 34  102023   1000     Windhoek    1    0   No    1  388
#> 35    1000  54903       Taipeh    0    0   No    0  660
#> 36    1000  20123     Windhoek    0    1   No    0  907
#> 37   54903   1000       Taipeh    1    0   No    0  390
#> 38    1000  20123     Windhoek    0    1   No    1  871
#> 39   54903  54903     Canberra    0    0   No    0  589
#> 40   20123   3502 Antananarivo    1    1  Yes    0  274
#> 41  102023  20123 Antananarivo    0    1   No    0  889
#> 42    3502 102023 Antananarivo    1    1  Yes    0  370
#> 43    3502  54903       Taipeh    1    0   No    1  255
#> 44    3502  54903     Windhoek    0    0  Yes    1  953
#> 45  102023  54903     Asuncion    1    1   No    1  724
#> 46   54903  54903     Canberra    1    1   No    1  214
#> 47    1000  54903       Taipeh    1    0  Yes    1  966
#> 48    3502 102023       Taipeh    0    1  Yes    1  154
#> 49  102023  54903     Asuncion    1    1   No    0   75
#> 50  102023  54903      Seattle    1    1  Yes    0  684
#> 51   54903 102023     Asuncion    0    1  Yes    1  390
#> 52   54903  54903     Windhoek    0    0   No    1  248
#> 53   54903 102023        Tokyo    0    0   No    0  583
#> 54   54903 102023       Taipeh    0    1   No    0  483
#> 55   20123  20123 Antananarivo    1    1   No    0  444
#> 56    3502 102023 Antananarivo    1    1  Yes    1  328
#> 57    1000  54903      Seattle    1    1  Yes    1  932
#> 58    1000 102023     Asuncion    1    1   No    0  950
#> 59    1000 102023    Ljubljana    0    1  Yes    0  328
#> 60   54903  54903     Asuncion    0    1  Yes    0   71
#> 61   54903   1000    Ljubljana    1    1  Yes    0   19
#> 62    1000   3502    Ljubljana    0    0   No    0  483
#> 63  102023   1000    Ljubljana    0    1  Yes    0  906
#> 64  102023   3502        Tokyo    1    1   No    1  271
#> 65    1000 102023    Ljubljana    1    1   No    0  508
#> 66   54903 102023       Taipeh    1    0  Yes    0  604
#> 67   20123  54903    Ljubljana    1    0  Yes    1  903
#> 68   54903   1000 Antananarivo    0    0   No    1  948
#> 69    1000  20123     Asuncion    1    0   No    1  836
#> 70  102023  20123       Taipeh    1    1  Yes    1  517
#> 71   20123   1000      Seattle    0    0   No    0  603
#> 72   54903   1000        Tokyo    1    1  Yes    1  627
#> 73  102023  54903     Asuncion    1    0  Yes    1  593
#> 74  102023   1000     Asuncion    1    1  Yes    1  355
#> 75    1000   1000    Ljubljana    1    0  Yes    1  397
#> 76   54903   1000 Antananarivo    0    0   No    1  393
#> 77  102023 102023     Windhoek    0    0   No    1  700
#> 78    1000  20123        Tokyo    0    0   No    1  186
#> 79  102023 102023        Tokyo    1    0   No    0  140
#> 80    1000  20123     Canberra    0    0  Yes    1  649
#> 81   20123  20123     Asuncion    1    1   No    1  743
#> 82  102023 102023        Tokyo    0    0   No    1   21
#> 83   54903  54903    Ljubljana    1    1  Yes    1  660
#> 84  102023  20123     Asuncion    1    1  Yes    1  740
#> 85    3502  20123 Antananarivo    0    1   No    0  894
#> 86   54903 102023     Windhoek    1    1   No    0  904
#> 87    3502   3502     Asuncion    0    0  Yes    1  900
#> 88   20123   1000        Tokyo    1    1  Yes    0  237
#> 89    3502 102023        Tokyo    1    0   No    0  648
#> 90    1000  20123 Antananarivo    0    0  Yes    1  182
#> 91  102023  20123 Antananarivo    0    1  Yes    1  531
#> 92    1000 102023     Windhoek    1    1   No    1  198
#> 93   54903  54903 Antananarivo    0    0   No    0  597
#> 94   54903 102023     Asuncion    0    1  Yes    0  799
#> 95  102023  54903     Windhoek    0    0   No    0  763
#> 96   20123   1000 Antananarivo    0    0  Yes    1  102
#> 97   20123   3502    Ljubljana    0    0   No    0  432
#> 98    3502   1000     Canberra    0    1   No    0  643
#> 99    3502  20123       Taipeh    1    1   No    0  435
#> 100  20123   1000       Taipeh    1    1  Yes    0  204